cpufreq: cppc: Add support for frequency invariance
The Frequency Invariance Engine (FIE) is providing a frequency scaling
correction factor that helps achieve more accurate load-tracking.

Normally, this scaling factor can be obtained directly with the help of
the cpufreq drivers as they know the exact frequency the hardware is
running at. But that isn't the case for CPPC driver.

Another way of obtaining that is using the AMU counter support, which is
already present in kernel, but that hardware is optional for platforms
and few of the platforms lack it.

This patch thus obtains this scaling factor using the existing logic
present in the cppc driver. Note that only one callback can be
registered using topology_set_scale_freq_tick() and if the one
implemented for AMU is registered first, then the CPPC one won't
register itself.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>